Documentation and related fixes
* rename to TurnPointersOnOff
* move TurnPointersOnOff to namespace
